tests: move tests for run-persistent helpers to the right module
Stranded in their original home by PR #1129.
